Issue #74 - Websocket used for display dashboards Rename Client into WebsocketClient (more readable) and switch it into websocket package
